packages/embedder/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
numpy>=1.26,<2.0
httpx==0.28.1
google-genai>=1.0.0
